moves grid list into separate component
